Refactor: move openapi utils into /util directory
