+++ src/lib/openapi/util/create-request-schema.ts
@@ -1,0 +1,17 @@
+import { OpenAPIV3 } from 'openapi-types';
+
+export const createRequestSchema = (
+    schemaName: string,
+): OpenAPIV3.RequestBodyObject => {
+    return {
+        description: schemaName,
+        required: true,
+        content: {
+            'application/json': {
+                schema: {
+                    $ref: `#/components/schemas/${schemaName}`,
+                },
+            },
+        },
+    };
+};
