+++ src/lib/openapi/util/create-response-schema.ts
@@ -1,0 +1,16 @@
+import { OpenAPIV3 } from 'openapi-types';
+
+export const createResponseSchema = (
+    schemaName: string,
+): OpenAPIV3.ResponseObject => {
+    return {
+        description: schemaName,
+        content: {
+            'application/json': {
+                schema: {
+                    $ref: `#/components/schemas/${schemaName}`,
+                },
+            },
+        },
+    };
+};
