+++ src/lib/routes/admin-api/state.ts
@@ -12,7 +12,8 @@ import { Logger } from '../../logger';
 import StateService from '../../services/state-service';
 import { IAuthRequest } from '../unleash-types';
 import { OpenApiService } from '../../services/openapi-service';
-import { createRequestSchema, createResponseSchema } from '../../openapi';
+import { createRequestSchema } from '../../openapi/util/create-request-schema';
+import { createResponseSchema } from '../../openapi/util/create-response-schema';
 import { emptyResponse } from '../../openapi/spec/empty-response';
 import { ExportParametersSchema } from '../../openapi/spec/export-parameters-schema';
