+++ src/lib/routes/admin-api/strategy.ts
@@ -14,7 +14,8 @@ import { Request, Response } from 'express';
 import { IAuthRequest } from '../unleash-types';
 import { OpenApiService } from '../../services/openapi-service';
 import { emptyResponse } from '../../openapi/spec/empty-response';
-import { createRequestSchema, createResponseSchema } from '../../openapi';
+import { createRequestSchema } from '../../openapi/util/create-request-schema';
+import { createResponseSchema } from '../../openapi/util/create-response-schema';
 import {
     strategySchema,
     StrategySchema,
