+++ src/lib/routes/admin-api/tag-type.ts
@@ -12,7 +12,8 @@ import { IUnleashServices } from '../../types/services';
 import TagTypeService from '../../services/tag-type-service';
 import { Logger } from '../../logger';
 import { IAuthRequest } from '../unleash-types';
-import { createRequestSchema, createResponseSchema } from '../../openapi';
+import { createRequestSchema } from '../../openapi/util/create-request-schema';
+import { createResponseSchema } from '../../openapi/util/create-response-schema';
 import { TagTypesSchema } from '../../openapi/spec/tag-types-schema';
 import { emptyResponse } from '../../openapi/spec/empty-response';
 import { ValidateTagTypeSchema } from '../../openapi/spec/validate-tag-type-schema';
