+++ src/lib/routes/admin-api/user-admin.ts
@@ -15,7 +15,8 @@ import { simpleAuthKey } from '../../types/settings/simple-auth-settings';
 import { anonymise } from '../../util/anonymise';
 import { OpenApiService } from '../../services/openapi-service';
 import { emptyResponse } from '../../openapi/spec/empty-response';
-import { createRequestSchema, createResponseSchema } from '../../openapi';
+import { createRequestSchema } from '../../openapi/util/create-request-schema';
+import { createResponseSchema } from '../../openapi/util/create-response-schema';
 import { userSchema, UserSchema } from '../../openapi/spec/user-schema';
 import { serializeDates } from '../../types/serialize-dates';
 import { usersSchema, UsersSchema } from '../../openapi/spec/users-schema';
