+++ src/lib/routes/admin-api/user.ts
@@ -9,7 +9,8 @@ import UserFeedbackService from '../../services/user-feedback-service';
 import UserSplashService from '../../services/user-splash-service';
 import { ADMIN, NONE } from '../../types/permissions';
 import { OpenApiService } from '../../services/openapi-service';
-import { createRequestSchema, createResponseSchema } from '../../openapi';
+import { createRequestSchema } from '../../openapi/util/create-request-schema';
+import { createResponseSchema } from '../../openapi/util/create-response-schema';
 import { emptyResponse } from '../../openapi/spec/empty-response';
 import { meSchema, MeSchema } from '../../openapi/spec/me-schema';
 import { serializeDates } from '../../types/serialize-dates';
