+++ src/lib/routes/auth/simple-password-provider.ts
@@ -7,7 +7,8 @@ import { IUnleashServices } from '../../types';
 import { NONE } from '../../types/permissions';
 import Controller from '../controller';
 import { IAuthRequest } from '../unleash-types';
-import { createRequestSchema, createResponseSchema } from '../../openapi';
+import { createRequestSchema } from '../../openapi/util/create-request-schema';
+import { createResponseSchema } from '../../openapi/util/create-response-schema';
 import { userSchema, UserSchema } from '../../openapi/spec/user-schema';
 import { LoginSchema } from '../../openapi/spec/login-schema';
 import { serializeDates } from '../../types/serialize-dates';
