+++ src/lib/routes/health-check.ts
@@ -7,7 +7,7 @@ import { OpenApiService } from '../services/openapi-service';
 
 import Controller from './controller';
 import { NONE } from '../types/permissions';
-import { createResponseSchema } from '../openapi';
+import { createResponseSchema } from '../openapi/util/create-response-schema';
 import { HealthCheckSchema } from '../openapi/spec/health-check-schema';
 
 export class HealthCheckController extends Controller {
